# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: OAuth 2.0 Lab API
  version: 1.0.0
  description: Unified API specification for all language implementations

servers:
  - url: http://localhost:8080
    description: Java Server
  - url: http://localhost:8000
    description: Python Server
  - url: http://localhost:3001
    description: Node.js Server
  - url: http://localhost:8888
    description: PHP Server
  - url: http://localhost:9000
    description: Go Server

paths:
  /oauth/authorize:
    get:
      summary: Authorization endpoint
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
        - name: scope
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        '302':
          description: Redirect to login or callback

  /oauth/token:
    post:
      summary: Token endpoint
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                grant_type:
                  type: string
                  enum: [authorization_code, client_credentials, refresh_token]
                code:
                  type: string
                client_id:
                  type: string
                client_secret:
                  type: string
      responses:
        '200':
          description: Token response
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                  expires_in:
                    type: integer